Remove unneeded abstraction
